Licensed software under GPL
This is very much NOT what the FSF recommends, as I do not have the 300 lines of code that they recommend a program should have to be licensed under the GPL. However, I chose to license this program under the GPL for the non-code assets, namely the databases besides every_word.txt, which comes from dwyl/english-words and is part of the public domain under The Unlicense. I know that some gematria sites (like gematrix.org) sell their own databases as csv files. I don't like the idea of people using my databases to use in propiertary software, so I decided to license the databases (besides every_word.txt) and the programs used to generate them (except the program to generate sentences.tar.gz, since the database used for that file is a non-trival size and would put me over well over Github's recommended repo size limit of 1 GB.) If you do consider each line in the databases as lines of code, then I'm well past the 300 line recommendation, so it's not that big of a deal, anyways.

Licensing under the GPL meant that I had to modify the programs that read from these databases and re-write them if they are removed for any reason, since the FSF says to add a notice in each file and this interferes with the programs. I also had to modify the main file to add an option to print out a disclaimer about the GPL, although this was optional. Finally, I added the license (under the appropriately named LICENSE file) so the user could read it at their leisure.

--Matteo Raso